@@ -1,24 +1,21 @@
 name: pending_operations
 description: A class for tracking operations and waiting until all of them have completed
-version: 1.0.1
-authors:
-  - Timm Preetz <timm.preetz@aboutyou.com>
-  - Henri Beck <henri.beck@aboutyou.com>
-homepage: https://github.com/aboutyou/dart_packages/tree/master/packages/state_queue
+version: 2.0.0-nullsafety.0
+homepage: https://github.com/aboutyou/dart_packages/tree/master/packages/pending_operations
 repository: https://github.com/aboutyou/dart_packages
 
 environment:
-  sdk: ">=2.3.0 <3.0.0"
-  flutter: ">=1.7.0 <2.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
+  flutter: ">=1.24.0-3.0.pre <2.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
 dev_dependencies:
-  test: any
-
-  pedantic: ^1.8.0+1
-
   flutter_test:
     sdk: flutter
+
+  pedantic: ^1.10.0-nullsafety.2
+
+  test: any
