@@ -8,8 +8,8 @@ import 'package:test/test.dart' as test;
 @isTest
 void blocTest<B extends StateQueue<State>, State>(
   String description, {
-  @required B Function() build,
-  FutureOr<void> Function(B bloc) act,
+  required B Function() build,
+  FutureOr<void> Function(B bloc)? act,
 
   /// Iterable of either `State`s or `Matcher`s (can be mixed)
   ///
@@ -17,7 +17,7 @@ void blocTest<B extends StateQueue<State>, State>(
   /// starting with the initial state.
   ///
   /// The length of emitted states must be equal to the length of this argument.
-  Iterable expect,
+  Iterable? expect,
 }) {
   test.test(description, () async {
     await runBlocTest<B, State>(
@@ -30,9 +30,9 @@ void blocTest<B extends StateQueue<State>, State>(
 
 @visibleForTesting
 Future<void> runBlocTest<B extends StateQueue<State>, State>({
-  @required B Function() build,
-  FutureOr<void> Function(B bloc) act,
-  Iterable expect,
+  required B Function() build,
+  FutureOr<void> Function(B bloc)? act,
+  Iterable? expect,
 }) async {
   final collectStates = expect != null;
 
