@@ -49,8 +49,7 @@ void main() {
       throwsA(
         // ignore: avoid_annotating_with_dynamic
         (dynamic e) =>
-            e is TestFailure &&
-            e.message.contains('Which: shorter than expected'),
+            e is TestFailure && e.message.contains('shorter than expected'),
       ),
     );
   });
@@ -69,8 +68,7 @@ void main() {
       throwsA(
         // ignore: avoid_annotating_with_dynamic
         (dynamic e) =>
-            e is TestFailure &&
-            e.message.contains('Which: longer than expected'),
+            e is TestFailure && e.message.contains('longer than expected'),
       ),
     );
   });
@@ -90,7 +88,7 @@ void main() {
         // ignore: avoid_annotating_with_dynamic
         (dynamic e) =>
             e is TestFailure &&
-            e.message.contains('Which: was <1> instead of <7> at location [1]'),
+            e.message.contains('at location [1] is <1> instead of <7>'),
       ),
     );
   });
